create shared-types package
