+++ apps/vue/package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@fmc/shared-styles": "*",
+    "@fmc/shared-types": "*",
     "vue": "^3.3.9",
     "vue-router": "^4.2.5"
   },
