+++ apps/vue/src/helpers/challenges.ts
@@ -1,11 +1,4 @@
-export interface Challenge {
-  title: string
-  link: string
-  difficulty: string
-  developer?: string
-  tags?: string[]
-  isNew?: boolean
-}
+import type { Challenge } from '@fmc/shared-types'
 
 export const challenges = new Map<string, Challenge>([
   [
