+++ apps/vue/src/helpers/contributors.ts
@@ -1,7 +1,4 @@
-interface Contributor {
-  name: string
-  pic: string
-}
+import type { Contributor } from '@fmc/shared-types'
 
 export const contributors = new Map<string, Contributor>([
   ['arpansaha13', { name: 'Arpan Saha', pic: 'https://avatars.githubusercontent.com/u/82361490' }],
