+++ shared/types/README.md
@@ -1,0 +1,3 @@
+# @fmc/shared-types
+
+A single source for all common types across all workspaces.
