+++ shared/types/index.d.ts
@@ -1,0 +1,14 @@
+export interface Challenge {
+  title: string;
+  link: string;
+  difficulty: string;
+  developer?: string;
+  contributors?: string[];
+  tags?: string[];
+  isNew?: boolean;
+}
+
+export interface Contributor {
+  name: string;
+  pic: string;
+}
